backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0"]
build-backend = "setuptools.build_meta"

[project]
name = "even-wear-shuffle"
version = "0.1.0"
description = "A fair-shuffle sync engine for Discogs collections."
authors = [
    {name = "Omar Ramos Escoto"}
]
dependencies = [
    "requests",
    "python-dotenv",
]
requires-python = ">=3.10"

[tool.setuptools.packages.find]
where = ["src"]